Hakkuna_Mattata_AI_Interview/resume_parser/requirements.txt
streamlit
PyMuPDF
python-dotenv
requests
